refactor: Update glioma_data_module.yaml and data_transformations.yaml
Refactor the glioma_data_module.yaml and data_transformations.yaml files to improve the organization and readability of the code. The changes include:
- Adding a new file glioma_data_module.yaml with updated configuration for the GliomaSegmentationDataModule.
- Adding a new file data_transformations.yaml with updated configuration for the data transformations.
- Removing the old file lightning_data_module.yaml.

These changes aim to enhance the modularity and maintainability of the codebase.
