+++ configs/data/data_transformations.yaml
@@ -1,0 +1,20 @@
+# data_transformations.yaml 
+
+train_transforms:
+  - _target_: monai.transforms.LoadImage
+    image_only: False
+  - _target_: monai.transforms.EnsureChannelFirst
+  - _target_: monai.transforms.ScaleIntensity
+  - _target_: monai.transforms.RandRotate
+    range_x: 0.1
+    prob: 0.5
+  - _target_: monai.transforms.RandFlip
+    prob: 0.5
+  - _target_: monai.transforms.ToTensor
+
+val_transforms:
+  - _target_: monai.transforms.LoadImage
+    image_only: False
+  - _target_: monai.transforms.EnsureChannelFirst
+  - _target_: monai.transforms.ScaleIntensity
+  - _target_: monai.transforms.ToTensor
