+++ configs/data/glioma_data_module.yaml
@@ -1,0 +1,14 @@
+# glioma_data_module.yaml
+
+_target_: glioma_segmentation.data.glioma_segmentation_data_module.GliomaSegmentationDataModule
+
+use_sample_data: True
+base_dir: ${hydra:runtime.cwd}
+
+train_dir: ${base_dir}/data_sample/training_data if ${.use_sample_data} else ${base_dir}/data/training_data
+val_dir: ${base_dir}/data_sample/validation_data if ${.use_sample_data} else ${base_dir}/data/validation_data
+
+batch_size: 1
+use_cache: False
+
+transformations: ${transforms.medical_transforms}
