--- configs/data/lightning_data_module.yaml
@@ -1,36 +1,0 @@
-_target_: glioma_segmentation.data.lightning_data_module.Medical3DDataModule
-
-# Boolean to choose between sample data and full data
-use_sample_data: True  # Set to False to use the full dataset
-
-# Base directory for the project (Hydra's current working directory)
-base_dir: ${hydra:runtime.cwd}
-
-# Directory paths for training and validation data
-train_dir: ${base_dir}/data_sample/training_data if ${.use_sample_data} else ${base_dir}/data/training_data
-val_dir: ${base_dir}/data_sample/validation_data if ${.use_sample_data} else ${base_dir}/data/validation_data
-
-# General parameters
-batch_size: 1               # Adjust according to your system's capabilities
-use_cache: False            # Set to True if you want to use MONAI's CacheDataset
-
-# Transformations parameters
-transformations:
-  train_transforms:
-    - _target_: monai.transforms.LoadImage
-      image_only: False
-    - _target_: monai.transforms.EnsureChannelFirst
-    - _target_: monai.transforms.ScaleIntensity
-    - _target_: monai.transforms.RandRotate
-      range_x: 0.1
-      prob: 0.5
-    - _target_: monai.transforms.RandFlip
-      prob: 0.5
-    - _target_: monai.transforms.ToTensor
-
-  val_transforms:
-    - _target_: monai.transforms.LoadImage
-      image_only: False
-    - _target_: monai.transforms.EnsureChannelFirst
-    - _target_: monai.transforms.ScaleIntensity
-    - _target_: monai.transforms.ToTensor
